+++ src/DfE.ExternalApplications.Application/Applications/Queries/GetApplicationsByTemplateQueryValidator.cs
@@ -11,5 +11,9 @@ internal class GetApplicationsByTemplateQueryValidator : AbstractValidator<GetAp
         RuleFor(x => x.TemplateId)
             .NotEmpty()
             .WithMessage("Template ID is required");
+
+        RuleFor(x => x.Search)
+            .SetValidator(new ApplicationListingSearchCriteriaValidator()!)
+            .When(x => x.Search is not null);
     }
 }