+++ src/DfE.ExternalApplications.Application/Applications/Queries/GetMyApplicationsQueryValidator.cs
@@ -4,17 +4,12 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("DfE.ExternalApplications.Application.Tests")]
 namespace DfE.ExternalApplications.Application.Applications.Queries;
 
-internal class GetApplicationsForUserByExternalProviderIdQueryValidator : AbstractValidator<GetApplicationsForUserByExternalProviderIdQuery>
+internal class GetMyApplicationsQueryValidator : AbstractValidator<GetMyApplicationsQuery>
 {
-    public GetApplicationsForUserByExternalProviderIdQueryValidator()
+    public GetMyApplicationsQueryValidator()
     {
-        RuleFor(x => x.ExternalProviderId)
-            .NotEmpty()
-            .WithMessage("External Provider ID is required");
-
         RuleFor(x => x.Search)
             .SetValidator(new ApplicationListingSearchCriteriaValidator()!)
             .When(x => x.Search is not null);
     }
 }
-