+++ CHANGELOG.md
@@ -79,6 +79,11 @@ All notable changes to this service will be documented in this file.
 ### Notes
 - Added new Role assignment endpoint
 - New endpoint for retrieving tenant-wide applications
+- Added RG Visits Test env appsettings
+
+## [1.3.1] - Public Beta
+### Notes
+- Added the ability to search for applications by reference on the dashboard
 
 ## [1.3.2] - Public Beta
 ### Notes
@@ -99,3 +104,7 @@ All notable changes to this service will be documented in this file.
 ## [1.3.6] - Public Beta
 ### Notes
 - Improved caching and invalidate cache and permissions after a change is made.
+
+## [1.3.7] - Public Beta
+### Notes
+- Added the ability to search for applications by different parameters on the dashboard